两个主体界面

+++ app/src/main/java/xx/yy/hou/lz/define/LeiXing.kt
@@ -3,7 +3,7 @@ package xx.yy.hou.lz.define
 import java.io.Serializable
 
 class LeiXing( // 0是根类型
-  var id: Int,
-  var parentId: Int = 0,
+  var id: Long,
+  var parentId: Long = 0L,
   var name: String
-): Serializable
+) : Serializable
